Funciones ordenadas en main.js
